* in_tkt_sky.c (krb5_get_in_tkt_with_skey): Change cast from         krb5_pointer to krb5_const_pointer to ensure const integrity of         parameter.
        * in_tkt_ktb.c (keytab_keyproc): Add const argument to cast of
        keyseed to struct keytab_keyproc_arg to maintain const status.

        * conv_princ.c (krb5_524_conv_principal): Cast argument to
        memcpy to size_t.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13636 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/krb/conv_princ.c
@@ -161,7 +161,7 @@ krb5_524_conv_principal(context, princ, name, inst, realm)
 		       c = strnchr(compo->data, '.', compo->length);
 		       if (!c || (c - compo->data) >= INST_SZ - 1)
 			   return KRB5_INVALID_PRINCIPAL;
-		       memcpy(inst, compo->data, c - compo->data);
+		       memcpy(inst, compo->data, (size_t) (c - compo->data));
 		       inst[c - compo->data] = '\0';
 		   }
 		   break;
